requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
efficientnet_pytorch>=0.7.1
albumentations>=1.3.0
opencv-python-headless>=4.7.0
streamlit>=1.24.0
numpy>=1.24.0
Pillow>=9.5.0
requests>=2.31.0